qa-chatbot/pyproject.toml
[project]
name = "qa-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
]
